# perf(gatsby-plugin-mdx): lazily fetch file nodes for remark plugins