Lazily fetch the list of `File` nodes. Before it would always generate the list of file nodes (which does a shallow copy, which is expensive at scale). Afterwards it only does this when the files are actually read, by adding a "getter" property.

Benchmark numbers on gabe-fs-mdx (no plugins), for just 10k pages:

```
info bootstrap finished - 110.731s
success run page queries - 176.578s - 10002/10002 56.64/s
info Done building in 297.347 sec
```
to
```
info bootstrap finished - 96.578s
success run page queries - 159.652s - 10002/10002 62.65/s
info Done building in 266.202 sec
```

(This snippet is used at sourcing time and as part of resolvers so run queries is also affected, so it saved roughly 2x 15s)