@pvdz 2020-11-10
Getters have been in JS since ES5 and are very safe to use from compat standpoint.

The real danger is that, potentially, plugins don't cache the access to `api.files` or something, meaning they'd do a repetitive fetch every time. But caching it is also not safe, for example it might break gatsby-dev when new files are sourced.

I think this is an acceptable middle ground, also because I'm not sure whether anything actually uses this at all. I couldn't find this property described in the docs either. Worst case this is a perf regression for a plugin that does use it. But it shouldn't break it.